feat(inventory): implement complete inventory transfer API flow (ship, receive, cancel, change-dest)

+++ src/main/java/com/zone/agri/dto/product/ProductRequest.java
@@ -25,7 +25,7 @@ public class ProductRequest {
         private BigDecimal costPrice;
         private BigDecimal price;
         private BigDecimal wholesalePrice;
-        private Long initialStock;
+        private Integer initialStock;
         private BigDecimal shippingWeight;
         private String image;
